+++ src/test/java/SavingTest.java
@@ -45,7 +45,7 @@ public class SavingTest {
         return account.getID();
     }
 
-    private int accountBalance(Account account) {
+    private double accountBalance(Account account) {
         return account.getBalance();
     }
 }
